refactor(remote): remove news related from data module
