--- data/src/main/java/com/anangkur/wallpaper/data/model/ArticleEntity.kt
@@ -1,13 +1,0 @@
-package com.anangkur.wallpaper.data.model
-
-data class ArticleEntity(
-    val id: Int = 0,
-    val author: String? = "",
-    val title: String? = "",
-    val description: String? = "",
-    val url: String? = "",
-    val urlToImage: String? = "",
-    val publishedAt: String? = "",
-    val content: String? = "",
-    val category: String? = ""
-)
