--- data/src/main/java/com/anangkur/wallpaper/data/repository/ArticleRemote.kt
@@ -1,13 +1,0 @@
-package com.anangkur.wallpaper.data.repository
-
-import com.anangkur.wallpaper.data.model.ArticleEntity
-
-interface ArticleRemote {
-    suspend fun getTopHeadlinesNews(
-        apiKey: String?,
-        country: String?,
-        category: String?,
-        sources: String?,
-        q: String?
-    ): List<ArticleEntity>
-}
