+++ data/src/main/java/com/anangkur/wallpaper/data/repository/LocalDataStore.kt
@@ -1,0 +1,5 @@
+package com.anangkur.wallpaper.data.repository
+
+interface LocalDataStore {
+    fun isExpired(): Boolean
+}
