+++ data/src/main/java/com/anangkur/wallpaper/data/repository/RemoteDataStore.kt
@@ -1,0 +1,3 @@
+package com.anangkur.wallpaper.data.repository
+
+interface RemoteDataStore
